feat: Add order number to email subject and body, display prices and total amount
- Updated send_email_confirmation function to include order number in the subject line and display it only once in the body.
- Updated post_order_food and post_order_book functions to ensure order number, prices, and total amount are included in the order details.
- Updated order_confirmation.html template to display order number, prices, and total amount.
- Fixed issue with quantity being treated as a string by converting it to an integer.
