+++ .gitignore
@@ -169,3 +169,7 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
+
+# macOS specific files
+.DS_Store
+static/images/.DS_Store
