+++ requirements.txt
@@ -1,4 +1,3 @@
-# filepath: /Users/markthomson/git/python/pythonccino-web/requirements.txt
 blinker==1.9.0
 click==8.1.8
 Flask==3.1.0
