+++ templates/add_book.html
@@ -1,4 +1,3 @@
-<!-- filepath: /Users/markthomson/git/python/pythonccino-web/templates/add_book.html -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
