+++ templates/add_food.html
@@ -1,4 +1,3 @@
-<!-- filepath: /Users/markthomson/git/python/pythonccino-web/templates/add_food.html -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
