+++ templates/home.html
@@ -1,4 +1,3 @@
-<!-- filepath: /Users/markthomson/git/python/pythonccino-web/templates/home.html -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
