+++ templates/order_book.html
@@ -4,23 +4,46 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Order Book</title>
+    <script>
+        function addItem() {
+            const itemContainer = document.getElementById('item-container');
+            const newItem = document.createElement('div');
+            newItem.innerHTML = `
+                <label for="book_title">Book Title:</label>
+                <select name="book_title" required>
+                    {% for item in book_menu %}
+                    <option value="{{ item.title }}">{{ item.title }} - £{{ item.price }}</option>
+                    {% endfor %}
+                </select><br>
+                <label for="quantity">Quantity:</label>
+                <input type="number" name="quantity" min="1" required><br>
+            `;
+            itemContainer.appendChild(newItem);
+        }
+    </script>
 </head>
 <body>
     <h1>Order Book</h1>
     <form action="/order_book" method="post">
         <label for="name">Your Name:</label>
         <input type="text" id="name" name="name" required><br>
-
-        <label for="book_title">Book Title:</label>
-        <select id="book_title" name="book_title" required>
-            {% for item in book_menu %}
-                <option value="{{ item.title }}">{{ item.title }} - £{{ item.price }}</option>
-            {% endfor %}
-        </select><br>
-
-        <label for="quantity">Quantity:</label>
-        <input type="number" id="quantity" name="quantity" min="1" required><br>
-
+        <label for="email">Your Email:</label>
+        <input type="email" id="email" name="email"><br>
+        <div id="item-container">
+            <div>
+                <label for="book_title">Book Title:</label>
+                <select name="book_title" required>
+                    {% for item in book_menu %}
+                    <option value="{{ item.title }}">{{ item.title }} - £{{ item.price }}</option>
+                    {% endfor %}
+                </select><br>
+                <label for="quantity">Quantity:</label>
+                <input type="number" name="quantity" min="1" required><br>
+            </div>
+        </div>
+        <button type="button" onclick="addItem()">Add Another Item</button><br>
+        <label for="email_confirmation">Would you like an email confirmation?</label>
+        <input type="checkbox" id="email_confirmation" name="email_confirmation"><br>
         <input type="submit" value="Order">
     </form>
     <h2>Menus</h2>
