+++ templates/order_food.html
@@ -4,6 +4,23 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Order Food</title>
+    <script>
+        function addItem() {
+            const itemContainer = document.getElementById('item-container');
+            const newItem = document.createElement('div');
+            newItem.innerHTML = `
+                <label for="food_item">Food Item:</label>
+                <select name="food_item" required>
+                    {% for item in food_menu %}
+                    <option value="{{ item.name }}">{{ item.name }} - {{item.description}} £{{ item.price }}</option>
+                    {% endfor %}
+                </select><br>
+                <label for="quantity">Quantity:</label>
+                <input type="number" name="quantity" min="1" required><br>
+            `;
+            itemContainer.appendChild(newItem);
+        }
+    </script>
 </head>
 <body>
     <h1>Order Food</h1>
@@ -12,14 +29,19 @@
         <input type="text" id="name" name="name" required><br>
         <label for="email">Your Email:</label>
         <input type="email" id="email" name="email"><br>
-        <label for="food_item">Food Item:</label>
-        <select id="food_item" name="food_item" required>
-            {% for item in food_menu %}
-            <option value="{{ item.name }}">{{ item.name }} - {{item.description}} £{{ item.price }}</option>
-            {% endfor %}
-        </select><br>
-        <label for="quantity">Quantity:</label>
-        <input type="number" id="quantity" name="quantity" min="1" required><br>
+        <div id="item-container">
+            <div>
+                <label for="food_item">Food Item:</label>
+                <select name="food_item" required>
+                    {% for item in food_menu %}
+                    <option value="{{ item.name }}">{{ item.name }} - {{item.description}} £{{ item.price }}</option>
+                    {% endfor %}
+                </select><br>
+                <label for="quantity">Quantity:</label>
+                <input type="number" name="quantity" min="1" required><br>
+            </div>
+        </div>
+        <button type="button" onclick="addItem()">Add Another Item</button><br>
         <label for="email_confirmation">Would you like an email confirmation?</label>
         <input type="checkbox" id="email_confirmation" name="email_confirmation"><br>
         <input type="submit" value="Order">
